Always pass groups memberships - delete api

+++ core/services.go
@@ -92,16 +92,26 @@ func (app *Application) updatePoll(user *model.User, poll model.Poll) (*model.Po
 }
 
 func (app *Application) deletePoll(user *model.User, id string) error {
-	poll, err := app.storage.GetPoll(user, id, true, nil)
+	//get the poll
+	groupMembership, err := app.groups.GetGroupsMembership(user.Token)
+	if err != nil {
+		return fmt.Errorf("error getting poll when delete - %s", err)
+	}
+	poll, err := app.storage.GetPoll(user, id, true, groupMembership)
 	if err != nil {
 		return err
 	}
+	if poll == nil {
+		return fmt.Errorf("poll not found")
+	}
 
+	//check permission
 	err = app.checkPollPermission(user, poll, "delete")
 	if err != nil {
 		return err
 	}
 
+	//delete the poll
 	err = app.storage.DeletePoll(user, id)
 	if err != nil {
 		return err
